feat(close): extract useCloseRegister hook and CloseHeader component
- Extract all state, effects, and handlers to useCloseRegister hook (199 lines)
- Extract breadcrumb + title to CloseHeader component (42 lines)
- CloseRegister.jsx reduced from 331 to 131 lines (thin orchestrator)
- Fix Register.jsx import path for GeneralModal (PR 2A bug)
- All files ≤200 lines, zero behavior change
