+++ .atl/.skill-registry.cache.json
@@ -1,0 +1,3 @@
+{
+  "fingerprint": "705541c03838e767027d794f18771328670bc359"
+}
